Split test file into smaller coherent test files. Add tests to get coverage over 90%. Update coverage badge script to update README and MKDocs index page
