+++ pyproject.toml
@@ -137,7 +137,7 @@ omit = [
 
 [tool.pytest.ini_options]
 minversion = "6.0"
-addopts = "-s -vvv --color=yes -n auto --cov-report=html --cov-report=term --cov-report=json --cov=src --cov-branch --cov-fail-under=89"
+addopts = "-s -vvv --color=yes -n auto --cov-report=html --cov-report=term --cov-report=json --cov=src --cov-branch --cov-fail-under=90"
 
 [tool.mypy]
 pretty = true
