--- tests/test_injinja.py
@@ -1,858 +1,0 @@
-"""Tests for the injinja module."""
-
-# Standard Library
-import json
-import pathlib
-import shutil
-import textwrap
-from unittest.mock import patch
-
-# Third Party
-import pytest
-import yaml
-from jinja2.filters import FILTERS
-
-# Our Libraries
-from injinja.injinja import (
-    ConfigSchemaValidationError,
-    JSONSchemaLoadingError,
-    PydanticConfigSchemaLoadingError,
-    _load_json_schema,
-    _load_pydantic_model,
-    dict_from_keyvalue_list,
-    dict_from_prefixes,
-    expand_files_list,
-    get_environment_variables,
-    get_functions,
-    load_config,
-    main,
-    map_env_to_confs,
-    merge,
-    merge_template,
-    parse_stdin_content,
-    read_env_file,
-    reduce_confs,
-    validate_config_with_jsonschema,
-    validate_config_with_pydantic,
-    validate_config_with_schema,
-)
-
-
-class TestEnvironmentVariables:
-    """Test environment variable handling functions."""
-
-    def test_dict_from_keyvalue_list(self):
-        """Test dict_from_keyvalue_list converts key=value strings."""
-        result = dict_from_keyvalue_list(["KEY1=value1", "KEY2=value2"])
-        assert result == {"KEY1": "value1", "KEY2": "value2"}
-
-    def test_dict_from_keyvalue_list_none(self):
-        """Test dict_from_keyvalue_list with None input."""
-        result = dict_from_keyvalue_list(None)
-        assert result is None
-
-    def test_read_env_file(self, tmp_path):
-        """Test read_env_file reads .env files correctly."""
-        env_file = tmp_path / ".env"
-        env_file.write_text("KEY1=value1\nKEY2=value2\n# Comment\nKEY3=value3")
-
-        result = read_env_file(str(env_file))
-        assert result == {"KEY1": "value1", "KEY2": "value2", "KEY3": "value3"}
-
-    def test_read_env_file_nonexistent(self):
-        """Test read_env_file returns None for nonexistent file."""
-        result = read_env_file("/nonexistent/file.env")
-        assert result is None
-
-    def test_dict_from_prefixes(self):
-        """Test dict_from_prefixes extracts prefixed env vars."""
-        with patch.dict("os.environ", {"PREFIX_VAR1": "value1", "PREFIX_VAR2": "value2", "OTHER": "value3"}):
-            result = dict_from_prefixes(["PREFIX_"])
-            assert "PREFIX_VAR1" in result
-            assert "PREFIX_VAR2" in result
-            assert "OTHER" not in result
-
-    def test_dict_from_prefixes_none(self):
-        """Test dict_from_prefixes with None input."""
-        result = dict_from_prefixes(None)
-        assert result is None
-
-    def test_get_environment_variables(self, tmp_path):
-        """Test get_environment_variables merges all sources."""
-        # Create a test .env file
-        env_file = tmp_path / ".env"
-        env_file.write_text("FILE_KEY=file_value")
-
-        with patch.dict("os.environ", {"PREFIX_KEY": "prefix_value"}):
-            result = get_environment_variables(
-                env_flags=["CLI_KEY=cli_value", str(env_file)], prefixes_list=["PREFIX_"]
-            )
-
-            assert result["CLI_KEY"] == "cli_value"
-            assert result["FILE_KEY"] == "file_value"
-            assert result["PREFIX_KEY"] == "prefix_value"
-
-
-class TestCustomFunctions:
-    """Test custom function loading."""
-
-    def test_get_functions(self, tmp_path):
-        """Test get_functions loads custom functions from Python files."""
-        func_file = tmp_path / "get_functions_functions.py"
-        func_file.write_text("""
-def test_custom():
-    return True
-
-def filter_uppercase(value):
-    return value.upper()
-
-def regular_function():
-    return "not included"
-""")
-
-        result = get_functions([str(func_file)])
-        assert "custom" in result["tests"]
-        assert "uppercase" in result["filters"]
-        assert "regular_function" not in result["tests"]
-        assert "regular_function" not in result["filters"]
-
-    def test_get_functions_empty(self):
-        """Test get_functions with no function files."""
-        result = get_functions([])
-        assert result == {"tests": {}, "filters": {}}
-
-
-class TestConfigLoading:
-    """Test configuration file loading functions."""
-
-    def test_load_config_json(self, tmp_path):
-        """Test load_config with JSON file."""
-        json_file = tmp_path / "config.json"
-        json_file.write_text('{"key": "value"}')
-
-        result = load_config(str(json_file))
-        assert result == {"key": "value"}
-
-    def test_load_config_yaml(self, tmp_path):
-        """Test load_config with YAML file."""
-        yaml_file = tmp_path / "config.yml"
-        yaml_file.write_text("key: value")
-
-        result = load_config(str(yaml_file))
-        assert result == {"key": "value"}
-
-    def test_load_config_toml(self, tmp_path):
-        """Test load_config with TOML file."""
-        toml_file = tmp_path / "config.toml"
-        toml_file.write_text('key = "value"')
-
-        result = load_config(str(toml_file))
-        assert result == {"key": "value"}
-
-    def test_load_config_with_template(self, tmp_path):
-        """Test load_config with Jinja2 templating."""
-        json_file = tmp_path / "config.json"
-        json_file.write_text('{"key": "{{ var }}"}')
-
-        result = load_config(str(json_file), {"var": "templated_value"})
-        assert result == {"key": "templated_value"}
-
-    def test_load_config_unsupported(self, tmp_path):
-        """Test load_config raises error for unsupported file type."""
-        txt_file = tmp_path / "config.txt"
-        txt_file.write_text("content")
-
-        with pytest.raises(ValueError, match="File type.*not supported"):
-            load_config(str(txt_file))
-
-    def test_parse_stdin_content(self):
-        """Test parse_stdin_content parses different formats."""
-        # Test JSON
-        result = parse_stdin_content('{"key": "value"}', "json")
-        assert result == {"key": "value"}
-
-        # Test YAML
-        result = parse_stdin_content("key: value", "yaml")
-        assert result == {"key": "value"}
-
-        # Test TOML
-        result = parse_stdin_content('key = "value"', "toml")
-        assert result == {"key": "value"}
-
-        # Test unsupported format
-        with pytest.raises(ValueError, match="Unsupported stdin format"):
-            parse_stdin_content("content", "xml")
-
-
-class TestTemplating:
-    """Test template merging functions."""
-
-    def test_merge_template_without_config(self, tmp_path):
-        """Test merge_template without configuration."""
-        template_file = tmp_path / "template.txt"
-        template_file.write_text("Hello World")
-
-        result = merge_template(str(template_file), None)
-        assert result == "Hello World"
-
-    def test_merge_template_with_config(self, tmp_path):
-        """Test merge_template with configuration."""
-        template_file = tmp_path / "template.txt"
-        template_file.write_text("Hello {{ name }}")
-
-        result = merge_template(str(template_file), {"name": "Test"})
-        assert result == "Hello Test"
-
-    # Commented out - Jinja2 version-specific behavior varies
-    # def test_merge_template_undefined_error(self, tmp_path):
-    #     """Test merge_template raises error for undefined variables in Jinja2 v3+."""
-    #     template_file = tmp_path / "template.txt"
-    #     template_file.write_text("Hello {{ undefined_var }}")
-    #
-    #     # Test with Jinja2 v3+ where StrictUndefined is used
-    #     import jinja2
-    #     if int(jinja2.__version__[0]) >= 3:
-    #         with pytest.raises(jinja2.exceptions.UndefinedError):
-    #             merge_template(str(template_file), {})
-    #     else:
-    #         # In older versions, it doesn't raise by default
-    #         result = merge_template(str(template_file), {})
-    #         assert "undefined_var" in result  # Variable remains in template
-
-
-class TestConfigMerging:
-    """Test configuration merging functions."""
-
-    def test_map_env_to_confs(self, tmp_path):
-        """Test map_env_to_confs loads and templates configs."""
-        config_file = tmp_path / "config.json"
-        config_file.write_text('{"key": "{{ env_var }}"}')
-
-        result = map_env_to_confs([str(config_file)], {"env_var": "value"})
-        assert result == [{"key": "value"}]
-
-    def test_reduce_confs(self):
-        """Test reduce_confs merges configurations."""
-        confs = [{"key1": "value1", "shared": "first"}, {"key2": "value2", "shared": "second"}]
-
-        result = reduce_confs(confs)
-        assert result["key1"] == "value1"
-        assert result["key2"] == "value2"
-        assert result["shared"] == "second"  # Later configs override
-
-
-class TestMergeFunction:
-    """Test the main merge function."""
-
-    def test_merge_basic(self, tmp_path):
-        """Test basic merge functionality."""
-        template = tmp_path / "template.txt"
-        template.write_text("Hello {{ name }}")
-
-        config = tmp_path / "config.json"
-        config.write_text('{"name": "World"}')
-
-        result, diff = merge(config=[str(config)], template=str(template))
-
-        assert result == "Hello World"
-        assert diff is None
-
-    def test_merge_with_validation(self, tmp_path):
-        """Test merge with validation file."""
-        template = tmp_path / "template.txt"
-        template.write_text("Hello World")
-
-        validate = tmp_path / "validate.txt"
-        validate.write_text("Hello World")
-
-        result, diff = merge(template=str(template), validate=str(validate))
-
-        assert result == "Hello World"
-        assert diff == ""  # No difference
-
-    def test_merge_output_to_file(self, tmp_path):
-        """Test merge outputs to file."""
-        template = tmp_path / "template.txt"
-        template.write_text("Content")
-
-        output = tmp_path / "output.txt"
-
-        merge(template=str(template), output=str(output))
-
-        assert output.read_text() == "Content"
-
-    def test_merge_config_json_output(self, tmp_path, capsys):
-        """Test merge with config-json output."""
-        config = tmp_path / "config.json"
-        config.write_text('{"key": "value"}')
-
-        merge(config=[str(config)], output="config-json")
-
-        captured = capsys.readouterr()
-        output = json.loads(captured.out)
-        assert output == {"key": "value"}
-
-    def test_merge_config_yaml_output(self, tmp_path, capsys):
-        """Test merge with config-yaml output."""
-        config = tmp_path / "config.json"
-        config.write_text('{"key": "value"}')
-
-        merge(config=[str(config)], output="config-yaml")
-
-        captured = capsys.readouterr()
-        output = yaml.safe_load(captured.out)
-        assert output == {"key": "value"}
-
-    @pytest.mark.skip(
-        reason="""
-        Getting inconsistent results across multiple runs. 
-        I suspect due to the nature of setting filters in Jinja is modifying a global function
-    """
-    )
-    def test_merge_with_functions(self, tmp_path):
-        """Test merge with custom functions."""
-        func_file = tmp_path / "merge_with_functions" / "merge_with_functions.py"
-        func_file.parent.mkdir(parents=True, exist_ok=True)
-        func_file.write_text(
-            textwrap.dedent("""
-        def filter_lowercase(value):
-            return value.lower()
-        """)
-        )
-
-        # Need non-empty config for Jinja templating environment to be active so that custom functions can be used.
-        config = tmp_path / "merge_with_functions" / "merge_with_functions_config.json"
-        config.write_text('{"name": "WORLD"}')
-
-        template = tmp_path / "merge_with_functions" / "merge_with_functions_template.txt"
-        template.write_text("{{ name | lowercase }}")
-
-        f = get_functions([str(func_file)])
-        assert "lowercase" in f["filters"]
-
-        # The jinja2 TESTS and FILTERS dictionary is a global variable we are tweaking here.
-        FILTERS.update(f["filters"])
-
-        result, _ = merge(template=str(template), config=[str(config)], functions=[str(func_file)])
-
-        assert "lowercase" in FILTERS
-        assert result == "world"
-
-    @patch("sys.stdin")
-    def test_merge_with_stdin(self, mock_stdin, tmp_path):
-        """Test merge with stdin input."""
-        mock_stdin.isatty.return_value = False
-        mock_stdin.read.return_value = '{"stdin_key": "stdin_value"}'
-
-        template = tmp_path / "template.txt"
-        template.write_text("{{ stdin_key }}")
-
-        result, _ = merge(template=str(template), stdin_format="json")
-
-        assert result == "stdin_value"
-
-
-class TestMainFunction:
-    """Test the main CLI entry point."""
-
-    def test_main_basic(self, tmp_path):
-        """Test main function with basic arguments."""
-        template = tmp_path / "template.txt"
-        template.write_text("Test")
-
-        with patch("injinja.injinja.merge") as mock_merge:
-            mock_merge.return_value = ("Test", None)
-            main(["-t", str(template)])
-
-            mock_merge.assert_called_once()
-            call_args = mock_merge.call_args[1]
-            assert call_args["template"] == str(template)
-
-    def test_main_with_all_args(self, tmp_path):
-        """Test main function with all arguments."""
-        template = tmp_path / "template.txt"
-        template.write_text("Test")
-
-        config = tmp_path / "config.json"
-        config.write_text('{"key": "value"}')
-
-        with patch("injinja.injinja.merge") as mock_merge:
-            mock_merge.return_value = ("Test", None)
-            main(
-                [
-                    "-t",
-                    str(template),
-                    "-c",
-                    str(config),
-                    "-e",
-                    "KEY=VALUE",
-                    "-p",
-                    "PREFIX_",
-                    "-o",
-                    "output.txt",
-                    "--stdin-format",
-                    "json",
-                ]
-            )
-
-            mock_merge.assert_called_once()
-            call_args = mock_merge.call_args[1]
-            assert call_args["template"] == str(template)
-            assert str(config) in call_args["config"]
-            assert "KEY=VALUE" in call_args["env"]
-            assert "PREFIX_" in call_args["prefix"]
-            assert call_args["output"] == "output.txt"
-            assert call_args["stdin_format"] == "json"
-
-
-class TestEdgeCases:
-    """Test edge cases and error conditions."""
-
-    def test_empty_config_list(self):
-        """Test with empty configuration list."""
-        result = reduce_confs([])
-        assert result == {}
-
-    def test_merge_creates_parent_directories(self, tmp_path):
-        """Test merge creates parent directories for output."""
-        template = tmp_path / "template.txt"
-        template.write_text("Content")
-
-        output = tmp_path / "nested" / "dir" / "output.txt"
-
-        merge(template=str(template), output=str(output))
-
-        assert output.exists()
-        assert output.read_text() == "Content"
-
-    def test_expand_files_list(self, tmp_path):
-        """Test expand files list."""
-
-        pwd = pathlib.Path.cwd()
-        relative_tmp_path = (pwd / "tmp").relative_to(pwd)
-        temp_file_1 = relative_tmp_path / "expand_files_list" / "_expand_files_list_1.txt"
-        temp_file_1.parent.mkdir(parents=True, exist_ok=True)
-        temp_file_1.write_text("Content")
-
-        temp_file_2 = relative_tmp_path / "expand_files_list" / "_expand_files_list_2.txt"
-        temp_file_2.write_text("Content")
-
-        glob_path_str = str(relative_tmp_path / "expand_files_list") + "/*.txt"
-
-        result = sorted(expand_files_list(glob_path_str))
-        shutil.rmtree(relative_tmp_path / "expand_files_list")
-        assert result == sorted([str(temp_file_1), str(temp_file_2)])
-
-
-class TestSchemaValidation:
-    """Test schema validation functionality."""
-
-    def test_pydantic_model_loading_success(self, tmp_path):
-        """Test successful Pydantic model loading."""
-        # Create a simple Pydantic model file
-        model_file = tmp_path / "test_models.py"
-        model_file.write_text(
-            textwrap.dedent("""
-            from pydantic import BaseModel
-
-            class TestModel(BaseModel):
-                name: str
-                age: int
-        """)
-        )
-
-        # Test loading the model
-        model_cls = _load_pydantic_model(f"{model_file}::TestModel")
-        assert model_cls.__name__ == "TestModel"
-
-        # Test creating instance
-        instance = model_cls(name="test", age=30)
-        assert instance.name == "test"
-        assert instance.age == 30
-
-    def test_pydantic_model_loading_missing_file(self):
-        """Test Pydantic model loading with missing file."""
-        with pytest.raises(PydanticConfigSchemaLoadingError) as exc_info:
-            _load_pydantic_model("/nonexistent/file.py::TestModel")
-        assert "Module file '/nonexistent/file.py' not found" in str(exc_info.value)
-
-    def test_pydantic_model_loading_missing_class(self, tmp_path):
-        """Test Pydantic model loading with missing class."""
-        model_file = tmp_path / "test_models.py"
-        model_file.write_text("from pydantic import BaseModel\n\nclass OtherModel(BaseModel): pass")
-
-        with pytest.raises(PydanticConfigSchemaLoadingError) as exc_info:
-            _load_pydantic_model(f"{model_file}::MissingModel")
-        assert "Class 'MissingModel' not found in" in str(exc_info.value)
-
-    def test_pydantic_model_loading_invalid_syntax(self, tmp_path):
-        """Test Pydantic model loading with invalid Python syntax."""
-        model_file = tmp_path / "test_models.py"
-        model_file.write_text("invalid python syntax !!!")
-
-        with pytest.raises(SyntaxError):
-            _load_pydantic_model(f"{model_file}::TestModel")
-
-    def test_json_schema_loading_success(self, tmp_path):
-        """Test successful JSON schema loading."""
-        schema_file = tmp_path / "test_schema.json"
-        schema_data = {
-            "type": "object",
-            "properties": {"name": {"type": "string"}, "age": {"type": "integer", "minimum": 0}},
-            "required": ["name", "age"],
-        }
-        schema_file.write_text(json.dumps(schema_data))
-
-        result = _load_json_schema(str(schema_file))
-        assert result == schema_data
-
-    def test_json_schema_loading_missing_file(self):
-        """Test JSON schema loading with missing file."""
-        with pytest.raises(JSONSchemaLoadingError) as exc_info:
-            _load_json_schema("/nonexistent/file.json")
-        assert "Schema file '/nonexistent/file.json' not found" in str(exc_info.value)
-
-    def test_json_schema_loading_invalid_json(self, tmp_path):
-        """Test JSON schema loading with invalid JSON."""
-        schema_file = tmp_path / "test_schema.json"
-        schema_file.write_text("invalid json content {{{")
-
-        with pytest.raises(JSONSchemaLoadingError) as exc_info:
-            _load_json_schema(str(schema_file))
-        assert "Invalid JSON in schema file" in str(exc_info.value)
-
-    def test_json_schema_loading_non_json_file(self, tmp_path):
-        """Test JSON schema loading with non-.json file."""
-        schema_file = tmp_path / "test_schema.yaml"
-        schema_file.write_text("name: test")
-
-        with pytest.raises(JSONSchemaLoadingError) as exc_info:
-            _load_json_schema(str(schema_file))
-        assert "JSON Schema must be a .json file" in str(exc_info.value)
-
-    def test_pydantic_validation_success(self, tmp_path):
-        """Test successful Pydantic validation."""
-        # Create a Pydantic model
-        model_file = tmp_path / "test_models.py"
-        model_file.write_text(
-            textwrap.dedent("""
-            from pydantic import BaseModel
-
-            class ConfigModel(BaseModel):
-                database_url: str
-                port: int
-                debug: bool = False
-        """)
-        )
-
-        # Valid configuration
-        config = {"database_url": "postgresql://localhost:5432/test", "port": 8080, "debug": True}
-
-        # Should not raise an exception
-        validate_config_with_pydantic(config, f"{model_file}::ConfigModel")
-
-    def test_pydantic_validation_failure(self, tmp_path):
-        """Test Pydantic validation failure."""
-        # Create a Pydantic model
-        model_file = tmp_path / "test_models.py"
-        model_file.write_text(
-            textwrap.dedent("""
-            from pydantic import BaseModel
-
-            class ConfigModel(BaseModel):
-                database_url: str
-                port: int
-                debug: bool = False
-        """)
-        )
-
-        # Invalid configuration (missing required field, wrong type)
-        config = {
-            "port": "not_an_integer",
-            "debug": True,
-            # missing database_url
-        }
-
-        with pytest.raises(ConfigSchemaValidationError) as exc_info:
-            validate_config_with_pydantic(config, f"{model_file}::ConfigModel")
-
-        error_msg = str(exc_info.value)
-        assert "Pydantic validation failed" in error_msg
-        assert "database_url" in error_msg or "port" in error_msg
-
-    def test_pydantic_validation_invalid_format(self):
-        """Test Pydantic validation with invalid schema specification format."""
-        config = {"test": "value"}
-
-        with pytest.raises(PydanticConfigSchemaLoadingError) as exc_info:
-            validate_config_with_pydantic(config, "invalid_format")
-        assert "Invalid format" in str(exc_info.value)
-        assert "Expected format: 'module.py::ModelClass'" in str(exc_info.value)
-
-    def test_json_schema_validation_success(self, tmp_path):
-        """Test successful JSON Schema validation."""
-        # Create a JSON schema
-        schema_file = tmp_path / "test_schema.json"
-        schema_data = {
-            "type": "object",
-            "properties": {
-                "name": {"type": "string"},
-                "age": {"type": "integer", "minimum": 0},
-                "email": {"type": "string", "format": "email"},
-            },
-            "required": ["name", "age"],
-        }
-        schema_file.write_text(json.dumps(schema_data))
-
-        # Valid configuration
-        config = {"name": "John Doe", "age": 30, "email": "john@example.com"}
-
-        # Should not raise an exception
-        validate_config_with_jsonschema(config, str(schema_file))
-
-    def test_json_schema_validation_failure(self, tmp_path):
-        """Test JSON Schema validation failure."""
-        # Create a JSON schema
-        schema_file = tmp_path / "test_schema.json"
-        schema_data = {
-            "type": "object",
-            "properties": {"name": {"type": "string"}, "age": {"type": "integer", "minimum": 0}},
-            "required": ["name", "age"],
-        }
-        schema_file.write_text(json.dumps(schema_data))
-
-        # Invalid configuration
-        config = {
-            "name": 123,  # Should be string
-            "age": -5,  # Should be >= 0
-            # Missing required fields handled by schema
-        }
-
-        with pytest.raises(ConfigSchemaValidationError) as exc_info:
-            validate_config_with_jsonschema(config, str(schema_file))
-
-        error_msg = str(exc_info.value)
-        assert "Schema validation failed" in error_msg
-
-    def test_unified_schema_validation_pydantic(self, tmp_path):
-        """Test unified schema validation interface with Pydantic."""
-        # Create a Pydantic model
-        model_file = tmp_path / "test_models.py"
-        model_file.write_text(
-            textwrap.dedent("""
-            from pydantic import BaseModel
-
-            class UnifiedModel(BaseModel):
-                service_name: str
-                replicas: int = 1
-        """)
-        )
-
-        config = {"service_name": "api", "replicas": 3}
-
-        # Should not raise an exception (contains ::)
-        validate_config_with_schema(config, f"{model_file}::UnifiedModel")
-
-    def test_unified_schema_validation_jsonschema(self, tmp_path):
-        """Test unified schema validation interface with JSON Schema."""
-        # Create a JSON schema
-        schema_file = tmp_path / "unified_schema.json"
-        schema_data = {
-            "type": "object",
-            "properties": {"service_name": {"type": "string"}, "replicas": {"type": "integer", "minimum": 1}},
-            "required": ["service_name"],
-        }
-        schema_file.write_text(json.dumps(schema_data))
-
-        config = {"service_name": "api", "replicas": 2}
-
-        # Should not raise an exception (no ::)
-        validate_config_with_schema(config, str(schema_file))
-
-    def test_merge_with_schema_validation_success(self, tmp_path):
-        """Test merge function with successful schema validation."""
-        # Create config and template files
-        config_file = tmp_path / "config.yaml"
-        config_file.write_text("name: test-app\nport: 8080")
-
-        template_file = tmp_path / "template.txt"
-        template_file.write_text("App: {{ name }} on port {{ port }}")
-
-        # Create JSON schema
-        schema_file = tmp_path / "schema.json"
-        schema_data = {
-            "type": "object",
-            "properties": {"name": {"type": "string"}, "port": {"type": "integer"}},
-            "required": ["name", "port"],
-        }
-        schema_file.write_text(json.dumps(schema_data))
-
-        output_file = tmp_path / "output.txt"
-
-        # Should not raise an exception
-        merge(config=[str(config_file)], template=str(template_file), output=str(output_file), schema=str(schema_file))
-
-        assert output_file.exists()
-        assert "App: test-app on port 8080" in output_file.read_text()
-
-    def test_merge_with_schema_validation_failure(self, tmp_path):
-        """Test merge function with schema validation failure."""
-        # Create invalid config file
-        config_file = tmp_path / "config.yaml"
-        config_file.write_text("name: test-app\nport: 'not-a-number'")
-
-        template_file = tmp_path / "template.txt"
-        template_file.write_text("App: {{ name }} on port {{ port }}")
-
-        # Create JSON schema that requires port to be integer
-        schema_file = tmp_path / "schema.json"
-        schema_data = {
-            "type": "object",
-            "properties": {"name": {"type": "string"}, "port": {"type": "integer"}},
-            "required": ["name", "port"],
-        }
-        schema_file.write_text(json.dumps(schema_data))
-
-        output_file = tmp_path / "output.txt"
-
-        # Should raise ConfigSchemaValidationError due to validation failure
-        with pytest.raises(ConfigSchemaValidationError) as exc_info:
-            merge(
-                config=[str(config_file)], template=str(template_file), output=str(output_file), schema=str(schema_file)
-            )
-        assert "Schema validation failed" in str(exc_info.value)
-        assert "port" in str(exc_info.value)
-        assert not output_file.exists()  # Should not create output on validation failure
-
-    def test_configuration_validation_error_custom_exception(self):
-        """Test ConfigSchemaValidationError can be raised and caught."""
-        error_msg = "Test validation error"
-
-        with pytest.raises(ConfigSchemaValidationError) as exc_info:
-            raise ConfigSchemaValidationError(error_msg)
-
-        assert str(exc_info.value) == error_msg
-
-    def test_schema_validation_with_none_schema(self, tmp_path):
-        """Test that no validation occurs when schema is None."""
-        # Create a config file with test data
-        config_file = tmp_path / "config.yaml"
-        config_file.write_text("test: value")
-
-        # Call merge without schema - should not raise any validation error
-        template_file = tmp_path / "template.txt"
-        template_file.write_text("Value: {{ test }}")
-
-        output_file = tmp_path / "output.txt"
-
-        # Should work without schema validation
-        merge(
-            config=[str(config_file)],
-            template=str(template_file),
-            output=str(output_file),
-            schema=None,  # No schema validation
-        )
-
-        assert output_file.exists()
-        assert "Value: value" in output_file.read_text()
-
-    def test_validate_config_with_schema_pydantic_vs_json(self, tmp_path):
-        """Test that the unified interface correctly routes to Pydantic vs JSON Schema."""
-        # Test JSON Schema path (no :: in schema)
-        schema_file = tmp_path / "test.json"
-        schema_data = {"type": "object", "properties": {"name": {"type": "string"}}, "required": ["name"]}
-        schema_file.write_text(json.dumps(schema_data))
-
-        config = {"name": "test"}
-
-        # This should route to JSON Schema validation
-        validate_config_with_schema(config, str(schema_file))
-
-        # Test Pydantic path (with :: in schema)
-        model_file = tmp_path / "models.py"
-        model_file.write_text(
-            textwrap.dedent("""
-            from pydantic import BaseModel
-            
-            class TestModel(BaseModel):
-                name: str
-        """)
-        )
-
-        # This should route to Pydantic validation
-        validate_config_with_schema(config, f"{model_file}::TestModel")
-
-    def test_debug_mode_activation(self):
-        """Test that debug mode can be activated via --debug flag."""
-        # Standard Library
-        import sys
-
-        # Our Libraries
-        from injinja import injinja
-
-        # Save original state
-        original_argv = sys.argv[:]
-        original_debug_mode = injinja.DEBUG_MODE
-
-        try:
-            # Test with --debug in argv
-            sys.argv = ["injinja", "--debug", "other", "args"]
-
-            # Reload the module to trigger the debug mode detection
-            # Since the module is already loaded, we'll just test the logic
-            test_argv = ["injinja", "--debug", "other", "args"]
-
-            # Simulate the debug mode detection logic
-            debug_mode_activated = "--debug" in test_argv
-            assert debug_mode_activated is True
-
-            # Test without --debug in argv
-            test_argv = ["injinja", "other", "args"]
-            debug_mode_activated = "--debug" in test_argv
-            assert debug_mode_activated is False
-
-        finally:
-            # Restore original state
-            sys.argv = original_argv
-            injinja.DEBUG_MODE = original_debug_mode
-
-    def test_cli_argument_factory_custom_short_flag(self):
-        """Test CLI argument factory with custom short flag."""
-        # This tests lines 147-148 for custom short flag usage
-        # Our Libraries
-        from injinja.injinja import CLI_CONFIG
-
-        # Create a test config with custom short flag
-        test_config = dict(CLI_CONFIG)  # Copy existing config
-        test_config["test_custom"] = {
-            "short_flag": "-x",  # Custom short flag
-            "help": "Test custom short flag",
-        }
-
-        # Remove schema to avoid conflicts in this test
-        if "schema" in test_config:
-            del test_config["schema"]
-
-        # Manually test the custom short flag logic
-        flag_kwargs = test_config["test_custom"]
-        if "short_flag" in flag_kwargs:
-            custom_short_flag = flag_kwargs.pop("short_flag")
-            if custom_short_flag:  # Test the custom short flag path
-                short_flag = custom_short_flag
-                use_short_flag = True
-                assert use_short_flag is True
-                assert short_flag == "-x"
-
-    def test_error_handling_edge_cases(self):
-        """Test error handling edge cases to improve coverage."""
-        # Test ConfigSchemaValidationError with different message types
-        error1 = ConfigSchemaValidationError("Simple message")
-        assert str(error1) == "Simple message"
-
-        error2 = ConfigSchemaValidationError("")  # Empty message
-        assert str(error2) == ""
-
-        # Test multiple error instantiation
-        errors = [ConfigSchemaValidationError(f"Error {i}") for i in range(3)]
-        assert len(errors) == 3
-        assert all(isinstance(e, ConfigSchemaValidationError) for e in errors)
